implemented redux for article handling
