+++ client/src/components/articles/Article.js
@@ -3,57 +3,134 @@ import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { deleteArticle } from '../../actions/articleActions';
+import moment from 'moment';
+
+import Wrapper from '../layout/Wrapper';
+import BtnRound from '../layout/BtnRound';
+
+const cardButtons = [
+  {
+    "id": 1,
+    "type": "link",
+    "datavalue": "read",
+    "title": "Read",
+    "icon": "fas fa-book-open"
+  },
+  {
+    "id": 2,
+    "type": "button",
+    "datavalue": "save",
+    "title": "Save",
+    "icon": "fas fa-thumbtack"
+  },
+  {
+    "id": 3,
+    "type": "button",
+    "datavalue": "delete",
+    "title": "Delete",
+    "icon": "far fa-trash-alt"
+  },
+  {
+    "id": 4,
+    "type": "button",
+    "datavalue": "comment",
+    "title": "Comment",
+    "icon": "far fa-comments fa-lg"
+  }
+];
 
 class Article extends Component {
   state = {
-    showArticleInfo: false
+    showArticleInfo: false,
+    cardButtons: cardButtons,
+    comments: []
   };
 
   onDeleteClick = id => {
     this.props.deleteArticle(id);
   };
 
+  // return in utc to convert the date from the offset provided to UTC
+  // these dates have no timezone
+  formatDate = (date) => moment.utc(date).format('MM/DD/YYYY');
+
   render() {
-    const { id, name, email, phone } = this.props.article;
     const { showArticleInfo } = this.state;
+    const {_id, title, date, url, author, summary, comments} = this.props.article;
 
     return (
       <div className="card card-body mb-3">
-        <h4>
-          {name}{' '}
-          <i
+
+        <h6 className="text-left">{title} 
+          <i 
+            className="fas fa-caret-down ml-2" 
+            style={{ cursor: 'pointer' }}
             onClick={() =>
               this.setState({
                 showArticleInfo: !this.state.showArticleInfo
               })
             }
-            className="fas fa-sort-down"
-            style={{ cursor: 'pointer' }}
-          />
-          <i
-            className="fas fa-times"
-            style={{ cursor: 'pointer', float: 'right', color: 'red' }}
-            onClick={this.onDeleteClick.bind(this, id)}
-          />
-          <Link to={`article/edit/${id}`}>
-            <i
-              className="fas fa-pencil-alt"
-              style={{
-                cursor: 'pointer',
-                float: 'right',
-                color: 'black',
-                marginRight: '1rem'
-              }}
-            />
+          ></i>
+          <i 
+            className="fas fa-times app-color-2" 
+            style={{cursor: 'pointer', float: 'right'}}
+            onClick={this.onDeleteClick.bind(this, _id)}
+          ></i>
+          <Link to={`api/articles/${_id}`}>
+            <i 
+              className="fas fa-pencil-alt app-color-4" 
+              style={{cursor: 'pointer', float: 'right', marginRight: '1rem'}}
+            ></i>
           </Link>
-        </h4>
+        </h6>
+
+        <small className="text-left text-muted">By {author}, {this.formatDate(date)}</small>
+
         {showArticleInfo ? (
-          <ul className="list-group">
-            <li className="list-group-item">Email: {email}</li>
-            <li className="list-group-item">Phone: {phone}</li>
-          </ul>
+          <div> 
+            <hr></hr>
+            <p className="card-text">{summary}</p>
+            <div className="mt-auto">
+              <p className="text-right mb-1">
+                <small><span id="comment-length-{id}">{comments.length}</span> Comments</small>
+              </p>
+              <Wrapper>
+                {this.state.cardButtons
+                  .filter(btn => btn.type === "link")
+                  .map(linkBtn => (
+                    <a 
+                      key={linkBtn.id}
+                      href={url} 
+                      className={`card-btn mt-auto app-border-color-1 bg-white app-color-1 `}
+                      target="_blank" 
+                      rel="noopener noreferrer"
+                      role="button"
+                    >
+                      <i className={`${linkBtn.icon} mt-3`} aria-hidden="true" style={{verticalAlign: 'middle'}} />                    
+                    </a>
+                ))}
+
+                {this.state.cardButtons
+                  .filter(btn => btn.type !== "link")
+                  .map(cardBtn => (
+                    <BtnRound
+                      id={cardBtn.id}
+                      key={cardBtn.id}
+                      data-value={cardBtn.datavalue}
+                      title={cardBtn.title}
+                      icon={cardBtn.icon}
+                      url= {url ? url : null}
+                      // onClick={this.props.handleBtnClick}
+                    />
+                ))}
+              </Wrapper>
+            </div>
+          </div>
         ) : null}
+
       </div>
+
+
     );
   }
 }
