--- client/src/components/layout/ArticleItem/index.js
@@ -1,1 +1,0 @@
-export { default } from "./ArticleItem";
