--- client/src/components/layout/Articles/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Articles";
