--- client/src/components/layout/SearchCard/SearchCard.css
@@ -1,15 +1,0 @@
-.card {
-  background-size: cover;
-  background-position: center;
-  background-clip: content-box;
-  border: 6px solid white;
-  /* height: 400px;
-  width: 400px; */
-  width: 100%;
-  height: auto;
-  box-shadow: 0 3px 6px rgba(0,0,0,0.16), 0 3px 6px rgba(0,0,0,0.23);
-  position: relative;
-  margin: 0;
-  margin-top: 30px;
-}
-
