--- client/src/components/layout/SearchCard/index.js
@@ -1,1 +1,0 @@
-export { default } from "./SearchCard";
