+++ client/src/components/pages/Home/Home.js
@@ -1,94 +1,34 @@
-// import React from "react";
 import React, { Component } from 'react';
-// import { Consumer } from '../../context';
-
-import Wrapper from "../../layout/Wrapper";
-import Articles from "../../layout/Articles";
+import Articles from '../../articles/Articles';
 import "./Home.css";
 
 class Home extends Component {
 
-  ///////////////////////////
-  // handle state
-  ///////////////////////////
-
-  // load the images from the json array
-  state = {
-    // articles: articles
-  };
-  
   render() {
     return (
       <React.Fragment>
-
         <div className="row text-center mx-auto mt-2">
-
+          {/* Profile */}
           <div id="profile-section" className="col-md-2 container">
             <h6 className="app-display-4">User Profile</h6>
 
           </div>
-
+          {/* News Feed */}
           <div id="newsfeed-section" className="col-md-3 home">
             <h6 className="app-display-4">Latest News</h6>
-
-            {/* <!-- 2 line heading card view --> */}
-            <Wrapper>
-              {/* <Consumer> */}
-                { (value) => (    
-                  <React.Fragment>
-                    {value.articles.map(articleItem => (
-                        <Articles
-                        id={articleItem._id}
-                        key={articleItem._id}
-                        title={articleItem.title}
-                        date={articleItem.date}
-                        url={articleItem.url}
-                        summary={articleItem.summary}
-                        author={articleItem.author}
-                        comments={articleItem.comments}
-                        />
-                    ))} 
-                  </React.Fragment>
-                ) }
-              {/* </Consumer> */}
-            </Wrapper>
-
-            {/* <!-- full card view --> */}
-            {/* <Wrapper>
-              <Consumer>
-                { (value) => (    
-                  <React.Fragment>
-                    {value.articles.map(articleItem => (
-                        <Card
-                          id={articleItem._id}
-                          key={articleItem._id}
-                          title={articleItem.title}
-                          date={articleItem.date}
-                          url={articleItem.url}
-                          summary={articleItem.summary}
-                          author={articleItem.author}
-                          comments={articleItem.comments}
-                          // onClick={ e => this.handleImageSelected(e, articleItem.id)}
-                        />
-                      ))}
-                  </React.Fragment>
-                ) }
-              </Consumer>
-            </Wrapper> */}
-
+            <Articles></Articles>
           </div>
-
+          {/* Blog */}
           <div id="blog-section" className="col-md-3 container">
             <h6 className="app-display-4">Posts</h6>
 
           </div>
-
+          {/* Chat */}
           <div id="chat-section" className="col-md-2 container">
             <h6 className="app-display-4">Trending/Chat</h6>
 
           </div>
         </div>
-
       </React.Fragment>
     )
   }
