--- client/src/components/pages/NewsFeed/NewsFeed.js
@@ -1,38 +1,0 @@
-import React, { Component } from 'react';
-// import Wrapper from "../Wrapper";
-// import ArticleItem from "../ArticleItem";
-import "./NewsFeed.css";
-// import articles from "./articlesTemp.json";
-
-class NewsFeed extends Component {
-
-  ///////////////////////////
-  // handle state
-  ///////////////////////////
-
-  // load the images from the json array
-  state = {
-    // articles: articles
-  };
-  
-  render() {
-    return (
-      <section className="news-feed">
-        NEWSFEED PAGE
-        {/* <Wrapper>
-          {this.state.articles.map(articleItem => (
-            <ArticleItem
-              id={articleItem.id}
-              key={articleItem.id}
-              title={articleItem.title}
-              date={articleItem.date}
-              url={articleItem.url}
-            />
-          ))}
-        </Wrapper> */}
-      </section>
-    )
-  }
-}
-
-export default NewsFeed;
