--- client/src/components/pages/NewsFeed/index.js
@@ -1,1 +1,0 @@
-export { default } from "./NewsFeed";
