+++ client/src/reducers/articleReducer.js
@@ -12,18 +12,22 @@ const initialState = {
 };
 
 export default function(state = initialState, action) {
-    switch (action.Type) {
+    console.log('action type is:  ' + action.type);
+    switch (action.type) {
         case GET_ARTICLES:
+            console.log('made it to GET_ARTICLES');
             return {
                 ...state,
                 articles: action.payload
             };
         case GET_ARTICLE:
+            console.log('made it to GET_ARTICLE');
             return {
                 ...state,
                 article: action.payload
             };
         case DELETE_ARTICLE:
+            console.log('made it to DELETE_ARTICLE');
             return {
                 ...state,
                 articles: state.articles.filter(
@@ -31,11 +35,13 @@ export default function(state = initialState, action) {
                 )
             };
         case ADD_ARTICLE:
+            console.log('made it to ADD_ARTICLE');
             return {
                 ...state,
                 articles: [action.payload, ...state.articles]
             };
         case UPDATE_ARTICLE:
+            console.log('made it to UPDATE_ARTICLE');
             return {
                 ...state,
                 articles: state.articles.map(
@@ -45,6 +51,7 @@ export default function(state = initialState, action) {
                 )
             };
         default:
+            console.log('Hit the default.');
             return state;
     }
 };
